refactor(DoctorCard): 🧑‍💻 + DoctorActions
create new component DoctorActions ->extract
CardActions from Info component
